Move interface to domain project - remove redundant static using for JsonSerializer, honestly no idea why this was here - update generated MartenDb models
