+++ Nulah.Up.IntegrationTests/Mocks/TestUpBankApi.cs
@@ -3,9 +3,7 @@ using Nulah.UpApi.Domain.Api;
 using Nulah.UpApi.Domain.Api.Accounts;
 using Nulah.UpApi.Domain.Api.Categories;
 using Nulah.UpApi.Domain.Api.Transactions;
-using Nulah.UpApi.Lib;
-using Nulah.UpApi.Lib.ApiModels;
-using Nulah.UpApi.Lib.ApiModels.Accounts;
+using Nulah.UpApi.Domain.Interfaces;
 
 namespace Nulah.Up.IntegrationTests.Mocks;
