+++ UpApi/Nulah.UpApi.Lib/Controllers/CategoryController.cs
@@ -1,6 +1,7 @@
 ﻿using Marten;
 using Microsoft.Extensions.Logging;
 using Nulah.UpApi.Domain.Api.Categories;
+using Nulah.UpApi.Domain.Interfaces;
 using Nulah.UpApi.Domain.Models;
 
 namespace Nulah.UpApi.Lib.Controllers;
